+++ src/Soluble/Japha/Bridge/Driver/ConnectionInterface.php
@@ -4,8 +4,15 @@ namespace Soluble\Japha\Bridge\Driver;
 
 interface ConnectionInterface
 {
+    /**
+     * Return internal client (driver specific)
+     * @return mixed
+     */
     public function getClient();
 
 
+    /**
+     * @return void
+     */
     public function connect();
 }